# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MartEase APP",
        "description": "Author: Vikas Chauhan",
        "version": "3.0.0"
    },
    "paths": {
...
        "/api/v1/user/get-my-profile": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Get User Profile",
                "description": "Retrieve the authenticated user's profile information.",
                "operationId": "getMyProfile",
                "responses": {
                    "200": {
                        "description": "User Profile Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "johndoe@example.com"
                                                        },
                                                        "phone_number": {
                                                            "type": "string",
                                                            "example": "1234567890"
                                                        },
                                                        "profile_image": {
                                                            "type": "string",
                                                            "example": "http://example.com/profile.jpg",
                                                            "nullable": true
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Profile Get Successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user/change-password": {
            "post": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Update User's password",
                "description": "Update User's password.",
                "operationId": "updatePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "old_password",
                                    "new_password",
                                    "confirm_password"
                                ],
                                "properties": {
                                    "old_password": {
                                        "description": "Old password",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "confirm_password": {
                                        "description": "Confirm password, must match new password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Password changed successfully"
                    },
                    "400": {
                        "description": "Old password does not match"
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "new_password": [
                                                    "The new password field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/log-out": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "/logout",
                "description": "/logout.",
                "operationId": "de017d5f67fd9c887f5986bb34c903a1",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Account logout Successfully."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Token Expired or Not Existed"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/delete-my-account": {
            "get": {
                "tags": [
                    "User Profile"
                ],
                "summary": "Delete user account",
                "description": "Deletes the authenticated user's account along with associated user addresses.",
                "operationId": "deleteMyAccount",
                "responses": {
                    "200": {
                        "description": "Account deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Account Deleted Successfully"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}